private data in lifeForm
I added a boolean that keeps track of whether or not they are holding something, and a pointer to said object. This will be used to transfer control of something from the level to a lifeform. In the case of enemies, some will start with a coin or something that they release upon their death