changing surroundings functionality
surroundings will now be handled by creating a vector of gridlocs for
all visible tile for the lifeform. GridLoc now contains a Space value as
weel. This is still being worked on, but it compiles. The nimkip moves
need to be modified to use this, I am working on testing it with the
broblubs.