new surroundings method working
The new method of getting the surroundings works. There old version
still exists but code that uses the old version needs updated. I tested
it by changing the broblub moveRandom function so it does work. I also
added a function to mover that takes the current location and the
location to move to and chooses the correct direction to move to.